lambdacontainer/processpdffunction/requirements.txt

boto3
pdfminer-six @ git+https://github.com/pdfminer/pdfminer.six@master
rtree
scipy